Move config/commands.js to chat-plugins
